ns: move synth pointers from resource.c to ns.c hsk_ns_onrecv()
Synth names are decoded and returned to the client without any
lookups or resource decoding taking place. Moving this logic
also means sharing some methods from resource.c that now have to be
declared in the header file.

This is the logic that resolves direct queries for synth records
e.g. `dig _fs0000g._synth. IN A`
which is not even their intended use case.

The logic to split a synth record up into authority (NS) and
additional (A/AAAA) records happens elsewhere.