resource: fixup: refactor all referral types to use common struct
NS, GLUE and SYNTH records are all processed by hsk_resource_to_ns()
and so they each need to share a common struct so the right data
can be pulled out of memory indiscriminately by the function.

In hindsight, this commit probably belongs as a fixup to one of the
earlier refactors where "uneeded" code was removed.